printf 100%

+++ src/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: jzampier <jzampier@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/07/15 12:22:45 by jzampier          #+#    #+#             */
-/*   Updated: 2026/07/15 19:05:14 by jzampier         ###   ########.fr       */
+/*   Updated: 2026/07/16 16:02:26 by jzampier         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@ int	ft_printf(const char *str, ...);
 int	do_check(char c, va_list *args);
 int	ft_putchar(char c);
 int	ft_putstr(char *str);
-int	ft_putnbr(int n);
+int	ft_putnbr(long n);
 int	ft_putptr(unsigned long ptr);
 int	ft_puthex(unsigned long n, char *base);
